+++ backend-api/app/controllers/inspectionWorker.js
@@ -12,9 +12,10 @@ const InspectLinks = db.inspected_links;
 const KnownSites = db.cybersquat_known_sites;
 const parse = require("parse-domains");
 const {
-  jaroWinklerDistance,
-  levenshteinDistance,
-} = require("./stringSimilarity");
+  checkTyposquattingBitsquatting,
+  checkLevelsquattingCombosquatting,
+} = require("./cybersquat.controller");
+const { logging, flagging } = require("./logging.controller");
 
 db.mongoose
   .set("strictQuery", true)
@@ -151,129 +152,19 @@ checkCybersquatting = async (url) => {
   logging(`checkCybersquatting= ~checkStrings | ${checkStrings}`);
 
   /* --------------- Check For Levelsquatting or Combosquatting --------------- */
+  const trademarks = await KnownSites.find({});
   const levelCombosquattingDetected = await checkLevelsquattingCombosquatting(
+    trademarks,
     parsedDomain.hostname
   );
   if (levelCombosquattingDetected) return;
 
   /* ---- Check for Typosquatting/Bitsquatting with string similarity algos --- */
   const typoBitsquattingDetected = await checkTyposquattingBitsquatting(
+    trademarks,
     checkStrings
   );
   if (typoBitsquattingDetected) return;
-};
-
-// /* --- Levelquatting or Combosquatting - detect direct usage of trademark --- */
-// checkLevelsquattingCombosquatting = async (checkStrings) => {
-//   console.log(checkStrings);
-//   const regex = checkStrings.map(function (k) {
-//     return new RegExp(k);
-//   });
-
-//   const trademarks = await KnownSites.find({
-//     keyword: { $in: regex },
-//   });
-
-//   if (trademarks) {
-//     var flags = trademarks[0].keyword;
-//     for (let i = 1; i < trademarks.length; i++) {
-//       flags += `, ${trademarks[i].keyword}`;
-//     }
-
-//     logging(
-//       `checkLevelsquattingCombosquatting= ~trademark | ${flags}`
-//     );
-
-//     flagging(
-//       `- Levelsquatting/Combosquatting Detected\n\t- Direct usage of trademark(s) {${flags}} found`
-//     );
-
-//     return true;
-//   }
-
-//   return false;
-// };
-
-/* --- Levelquatting or Combosquatting - detect direct usage of trademark --- */
-checkLevelsquattingCombosquatting = async (parsedHostname) => {
-  const trademarks = await KnownSites.find({});
-
-  // remove any "-" or "." to account for cases like dh-s.bank, or db.s.bank
-  parsedHostname = parsedHostname.replace(/\-/g, "").replace(/\./g, "");
-
-  var flags = "";
-  for (let i = 0; i < trademarks.length; i++) {
-    if (parsedHostname.includes(trademarks[i].keyword)) {
-      flags += ` ${trademarks[i].keyword}`;
-    }
-  }
-
-  if (flags != "") {
-    logging(`checkLevelsquattingCombosquatting= ~trademark |${flags}`);
-
-    flagging(
-      `- Levelsquatting/Combosquatting Detected\n\t- Direct usage of trademark(s) {${flags} } found`
-    );
-
-    return true;
-  }
-
-  return false;
-};
-
-/* --- Typosquatting or Bitsquatting - detect indirect usage of trademark --- */
-checkTyposquattingBitsquatting = async (checkStrings) => {
-  var keywords = await KnownSites.find({});
-  keywords = keywords.map((record) => record.keyword);
-
-  var flags = "";
-
-  for (let i = 0; i < checkStrings.length; i++) {
-    for (let j = 0; j < keywords.length; j++) {
-      var jaroWinklerSimilarity = jaroWinklerDistance(
-        checkStrings[i],
-        keywords[j]
-      );
-      logging(
-        `checkTyposquattingBitsquatting= ~jaroWinklerSimilarity | Comparing ${checkStrings[i]} with ${keywords[j]}: ${jaroWinklerSimilarity}`
-      );
-
-      if (parseFloat(jaroWinklerSimilarity) >= 0.75) {
-        if (flags.length != 0) flags += "\n";
-        flags += `- Typosquatting/Bitsquatting Detected with Jaro-Winkler Algorithm\n\t- Similarity of {${checkStrings[i]}} with trademark {${keywords[j]}} is ${jaroWinklerSimilarity}`;
-      }
-
-      if (jaroWinklerSimilarity > 0.6) {
-        var levenshteinDistSimilarity = levenshteinDistance(
-          checkStrings[i],
-          keywords[j]
-        );
-        logging(
-          `checkTyposquattingBitsquatting= ~levenshteinDistSimilarity | Comparing ${checkStrings[i]} with ${keywords[j]}: ${levenshteinDistSimilarity}`
-        );
-
-        if (levenshteinDistSimilarity / checkStrings[i].length <= 1 / 3) {
-          if (flags.length != 0) flags += "\n";
-          flags += `- Typosquatting/Bitsquatting Detected with Levenshtein Distance\n\t- Distance of {${checkStrings[i]}} with trademark {${keywords[j]}} is ${levenshteinDistSimilarity}`;
-        }
-      }
-    }
-  }
-
-  if (flags.length != 0) {
-    flagging(flags);
-    return true;
-  }
-
-  return false;
-};
-
-logging = (message) => {
-  parentPort.postMessage(["log", message]);
-};
-
-flagging = (message) => {
-  parentPort.postMessage(["flag", message]);
 };
 
 terminatingWorker = (inspectedLink) => {
