+++ backend-api/app/controllers/logging.controller.js
@@ -1,0 +1,9 @@
+const { parentPort } = require("worker_threads");
+
+exports.logging = (message) => {
+  parentPort.postMessage(["log", message]);
+};
+
+exports.flagging = (message) => {
+  parentPort.postMessage(["flag", message]);
+};
